src/pyproject.toml
[project]
name = "833s-guardian"
version = "1.0.0a1"
description = "833's Guardian - Discord bot with paced queues, delays, caching, and a clean, maintainable architecture."
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.6.0",
  "python-dotenv>=1.0.1",
  "aiosqlite>=0.20.0",
  "aiohttp>=3.9.0",
  "psutil>=5.9.0",
]

[project.scripts]
guardian = "guardian.__main__:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]


[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["guardian*"]
